take-home/django/requirements.txt
Django==2.2.19
djangorestframework==3.10.2
django-cors-headers==3.1.0
django-cors-middleware==1.5.0
